cluster/cassis: Save gene IDs as array, not string
Store ID(s) of genes, which contribute to a promoter, as an array and
not a string (function get_all_promoters).

Implement function get_promoter_id for situations where a string is
required: It returns "id" if one gene contributed and return "id1+id2"
if two genes contributed. This will be used quite often by the CASSIS
plugin.