cluster/cassis: Calculate cluster length
Information about the length of the cluster prediction was missing.
Hence, after calling find_islands and sort_by_abundance, add code to the
detect function to find indices of first and last gene of a cluster. Do
this in respect to (i) all genes and (ii) all promoters (promoter
sequences). Save and print the length of the cluster prediction in terms
of genes and promoters.