cluster/cassis: Raise exception if invalid promoter
While getting the promoter sequences: Raise an user defined exception
if the location of two neighbouring genes is unclear
(InvalidLocationError) or if duplicate promoter sequences (two identical
promoter IDs, DuplicatePromoterError) where generated. Catch the
exception in the detect method, print an error message and skip the
CASSIS cluster detection for this SeqRecord.