cluster/cassis: Fix reported motif
While computing the cluster predictions based on the binding site
islands, incorrect motifs and motif scores where associated with the
predictions. This happened if there was more than one motif for a
prediction to choose from. Erroneously, the last (element in array)
matching motif was reported. The correct behavior is to report the motif
with the lowest (best) score.